requirements.txt
protobuf
fastmcp
paramiko
requests
cryptography
pydantic
python-dotenv
PyJWT
typing-extensions
google-api-python-client
google-auth
pandas
psycopg2
sqlalchemy

pytest
pytest-asyncio